+++ src/extension/features/accounts/clear-assistant/clearAssistantUtils.js
