+++ src/extension/features/accounts/clear-assistant/settings.js
