+++ src/extension/features/accounts/clear-assistant/styles.scss
@@ -10,6 +10,6 @@
   font-weight: 400;
 }
 
-#tk-auto-reconcile-container {
+#tk-clear-assistant-container {
   float: right;
 }
